+++ internal/config/defaults.go
@@ -26,10 +26,11 @@ func Default() *Configuration {
 			Runtime: "/run/cue",
 		},
 		Display: Display{
-			Server:          ServerXorg,
-			Number:          0,
-			VirtualTerminal: 2,
-			Cursor:          false,
+			Server:            ServerXorg,
+			Number:            0,
+			VirtualTerminal:   2,
+			Cursor:            CursorAuto,
+			CursorIdleTimeout: Duration(3 * time.Second),
 			Outputs: []Output{
 				{Name: "*", Mode: ModePreferred, Position: "0x0", Rotate: "normal"},
 			},